requirements.txt
future
zope.interface>=4.6.0
numpy
vtk
vtkAtamai
wxPython
